Simple_Real_Estate_Agent/pyproject.toml
[project]
name = "simple-real-estate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
]
